requirements.txt
chromadb
google-genai
python-dotenv
feedparser
numpy
fastapi
uvicorn
redis
pydantic
requests
